pyproject.toml
[tool.coverage.run]
branch = true
omit = ["config.py", "config-3.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
]

[tool.mypy]
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = 'generated.*'
ignore_errors = true

[[tool.mypy.overrides]]
module = 'choreo'
ignore_missing_imports = true
follow_imports = "skip"

[tool.pyright]
exclude = [
    "ctre_sim",
    "logs",
    # Default excludes
    "**/__pycache__",
    "**/.*",
]

[tool.pytest.ini_options]
addopts = "--strict-markers -v --maxfail=2"
pythonpath = "."
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
target-version = "py311"
preview = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # pyflakes
    "F",
    # flake8-bugbear
    "B",
    # isort
    "I",
    # pyupgrade
    "UP",
    # flake8-comprehensions
    "C4",
    # flake8-logging-format
    "G",
    # flake8-simplify
    "SIM",
    # flake8-print
    "T20",
    # perflint
    "PERF",
]
ignore = [
    "E501", # long lines
    "B905", # zip() without strict=
    "I001", # Unsorted imports
]

[tool.uv]
dev-dependencies = [
    "hypothesis>=6.112.1",
    "mypy>=1.8.0",
    "pytest>=7.2.0",
    "pytest-integration>=0.2.3",
]
prerelease = "allow"

[project]
name = "Reefscape2025"
version = "0.0.0"
description = "The TriSoncis 2025 Robot Code"
authors = [{ name = "The TriSonids", email = "contact@allendalerobotics.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<3.14"

dependencies = [
    "numpy~=2.1",
    "phoenix6~=25.2.2",
    "robotpy[apriltag]==2025.2.1.2",
    "robotpy-wpilib-utilities==2025.0.0",
    "photonlibpy==2025.1.1",
    "sleipnirgroup-choreolib>=2025.0.2",
    "robotpy-pathplannerlib",
    "pip",
]

[tool.robotpy]
requires = [
    "numpy~=2.1",
    "phoenix6~=25.2.2",
    "robotpy-wpilib-utilities==2025.0.0",
    "photonlibpy==2025.1.1",
    "sleipnirgroup-choreolib>=2025.0.2",
    "robotpy-pathplannerlib",
]
robotpy_version = "2025.2.1.2"
robotpy_extras = ["apriltag"]

[tool.robotpy.pyfrc]
isolated = true
